chore(FloatingLabel, LineRipple): Making resuasble component
