+++ src/FloatingLabel/floating-label.spec.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import { FloatingLabel } from './';
+
+describe('FloatingLabel', () => {
+  it('renders', () => {
+    mount(<FloatingLabel />);
+  });
+});
