+++ src/LineRipple/index.js
@@ -1,0 +1,8 @@
+import { simpleTag } from '../Base';
+
+export const LineRipple = simpleTag({
+  displayName: 'LineRipple',
+  classNames: 'mdc-line-ripple'
+});
+
+export default LineRipple;
