+++ src/LineRipple/line-ripple.spec.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import { LineRipple } from './';
+
+describe('LineRipple', () => {
+  it('renders', () => {
+    mount(<LineRipple />);
+  });
+});
