+++ src/rmwc.js
@@ -8,11 +8,13 @@ export * from './Drawer';
 export * from './Dialog';
 export * from './Elevation';
 export * from './Fab';
+export * from './FloatingLabel';
 export * from './FormField';
 export * from './GridList';
 export * from './Grid';
 export * from './Icon';
 export * from './IconToggle';
+export * from './LineRipple';
 export * from './LinearProgress';
 export * from './List';
 export * from './Menu';
